+++ library/LibUtilities/LinearAlgebra/NekLinSysIterGMRESLoc.h
@@ -106,8 +106,7 @@ private:
 
     // Arnoldi process
     void DoArnoldi(const int starttem, const int endtem, const int nLocal,
-                   // V_total(:,1:nd) total search directions
-                   Array<OneD, Array<OneD, NekDouble>> &V_local,
+                   Array<OneD, NekDouble> &w, Array<OneD, NekDouble> &wk,
                    // V[nd] current search direction
                    Array<OneD, NekDouble> &V1,
                    // V[nd+1] new search direction
@@ -130,6 +129,13 @@ private:
 
     static std::string lookupIds[];
     static std::string def;
+
+    // Hessenburg matrix
+    Array<OneD, Array<OneD, NekDouble>> m_hes;
+    // Hesseburg matrix after rotation
+    Array<OneD, Array<OneD, NekDouble>> m_Upper;
+    // Total search directions
+    Array<OneD, Array<OneD, NekDouble>> m_V_total;
 };
 } // namespace Nektar::LibUtilities